Evoastra Venture Pvt Ltd/requirements (1).txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
nltk>=3.8.0
tqdm>=4.65.0
pandas>=2.0.0

